move functions to avoid circular import
