+++ spyder/utils/programs.py
@@ -29,11 +29,15 @@ import time
 import psutil
 
 # Local imports
-from spyder.config.base import is_stable_version, running_under_pytest
+from spyder.config.base import (is_stable_version, running_under_pytest,
+                                get_home_dir, running_in_mac_app)
 from spyder.config.utils import is_anaconda
 from spyder.py3compat import PY2, is_text_string, to_text_string
 from spyder.utils import encoding
 from spyder.utils.misc import get_python_executable
+
+
+WINDOWS = os.name == 'nt'
 
 
 class ProgramError(Exception):
@@ -989,3 +993,68 @@ def is_spyder_process(pid):
         return any(conditions)
     except (psutil.NoSuchProcess, psutil.AccessDenied):
         return False
+
+
+def get_pyenv_path(name):
+    """Return the complete path of the pyenv."""
+    home = get_home_dir()
+    if WINDOWS:
+        path = osp.join(
+            home, '.pyenv', 'pyenv-win', 'versions', name, 'python')
+    elif name == '':
+        path = osp.join(home, '.pyenv', 'shims', 'python')
+    else:
+        path = osp.join(home, '.pyenv', 'versions', name, 'bin', 'python')
+    return path
+
+
+def get_list_pyenv_envs():
+    """Return the list of all pyenv envs found in the system."""
+    pyenv = 'pyenv'
+    if running_in_mac_app():
+        old_path = os.environ['PATH']
+        os.environ['PATH'] = os.pathsep.join([old_path, '/usr/local/bin'])
+        pyenv_path = find_program('pyenv')
+        os.environ['PATH'] = old_path  # restore PATH
+        if pyenv_path:
+            pyenv = pyenv_path
+    try:
+        out, err = subprocess.Popen(
+            [pyenv, 'versions', '--bare', '--skip-aliases'],
+            stdout=subprocess.PIPE,
+            stderr=subprocess.PIPE
+        ).communicate()
+        out = out.decode()
+        err = err.decode()
+    except Exception:
+        out = ''
+        err = ''
+    out = out.split('\n')
+    env_list = {}
+    for env in out:
+        data = env.split('/')
+        path = get_pyenv_path(data[-1])
+        if data[-1] == '':
+            name = 'internal' if running_in_mac_app(path) else 'system'
+        else:
+            name = 'pyenv: {}'.format(data[-1])
+        version = (
+            'Python 2.7' if data[-1] == '' else 'Python {}'.format(data[0]))
+        env_list[name] = (path, version)
+    return env_list
+
+
+def get_interpreter_info(path):
+    """Return version information of the selected Python interpreter."""
+    try:
+        out, err = subprocess.Popen(
+                [path, '-V'],
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE
+        ).communicate()
+        out = out.decode()
+        err = err.decode()
+    except Exception:
+        out = ''
+        err = ''
+    return out.strip()
