+++ spyder/widgets/status.py
@@ -20,7 +20,7 @@ from qtpy.QtWidgets import QHBoxLayout, QLabel, QMenu, QWidget
 # Local imports
 from spyder.config.base import _
 from spyder.utils.conda import get_list_conda_envs
-from spyder.utils.misc import get_list_pyenv_envs, get_interpreter_info
+from spyder.utils.programs import get_list_pyenv_envs, get_interpreter_info
 from spyder.utils.qthelpers import (add_actions, create_action,
                                     create_waitspinner)
